Removed Star Component and used existing material UI for star rating
