+++ package.json
@@ -11,9 +11,11 @@
     "@types/react-dom": "16.8.4",
     "@types/react-router-dom": "^4.3.3",
     "@types/relay-runtime": "^1.3.12",
+    "material-ui-rating": "^4.1.1",
     "nullthrows": "^1.1.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-rater": "^5.1.0",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1",
     "relay-runtime": "^4.0.0",
