--- src/book-ratings/RateButton.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-
-export default class RateButton extends React.Component 
-{
-    render()
-    {
-        return(
-            <div>
-                <p><h1>Book rating & commenting</h1></p>
-              
-            </div> )
-    }
-
-    //custom functions 
-}
