--- src/book-ratings/StarRateComponent.css
@@ -1,40 +1,0 @@
-:root
-{
-  --star_left: 20%;
-  --star_top: 75%;
-}
-
-
-.star1, .star2, .star3, .star4, .star5
-{
-  left: var(--star_left);
-  top: var(--star_top);
-
-  display: inline-block;
-  padding: 3px;
-  vertical-align: middle;
-  line-height: 1;
-  font-size: 1.5em;
-  color: grey;
-  transition: color .2s ease-out;
-}
-
-.star2 { left: 22%;}
-.star3 { left: 24%;}
-.star4 { left: 26%;}
-.star5 { left: 28%;}
-
-.star1:hover,.star2:hover,.star3:hover,.star4:hover,.star5:hover
-{
-  cursor: pointer;
-  color: gold;
-}
-.star1.is-selected , .star2.is-selected
-{
-  color: gold;
-}
-.star1.is-disabled:hover, .star2.star1.is-disabled:hover
-{
-  cursor: default;
-}
-
